Moved most of the eventing code into separate observable package
