--- core-observable/src/main/java/com/excelian/mache/MapEventListener.java
@@ -1,7 +1,0 @@
-package com.excelian.mache;
-
-import com.excelian.mache.coordination.CoordinationEntryEvent;
-
-public interface MapEventListener {
-    void send(CoordinationEntryEvent<?> event);
-}
