+++ core-observable/src/main/java/com/excelian/mache/events/BaseCoordinationEntryEventProducer.java
@@ -1,7 +1,7 @@
 package com.excelian.mache.events;
 
-import com.excelian.mache.MapEventListener;
-import com.excelian.mache.coordination.CoordinationEntryEvent;
+import com.excelian.mache.observable.MapEventListener;
+import com.excelian.mache.observable.coordination.CoordinationEntryEvent;
 
 public abstract class BaseCoordinationEntryEventProducer implements MapEventListener {
