+++ core-observable/src/main/java/com/excelian/mache/events/EventProducer.java
@@ -1,7 +1,7 @@
 package com.excelian.mache.events;
 
 
-import com.excelian.mache.coordination.CoordinationEntryEvent;
+import com.excelian.mache.observable.coordination.CoordinationEntryEvent;
 
 import java.util.concurrent.BlockingQueue;
