+++ core-observable/src/main/java/com/excelian/mache/events/integration/ActiveMQEventConsumer.java
@@ -1,6 +1,6 @@
 package com.excelian.mache.events.integration;
 
-import com.excelian.mache.coordination.CoordinationEntryEvent;
+import com.excelian.mache.observable.coordination.CoordinationEntryEvent;
 import com.excelian.mache.events.BaseCoordinationEntryEventConsumer;
 import com.google.gson.Gson;
 import org.slf4j.Logger;
