+++ core-observable/src/main/java/com/excelian/mache/events/integration/ActiveMQEventProducer.java
@@ -1,6 +1,6 @@
 package com.excelian.mache.events.integration;
 
-import com.excelian.mache.coordination.CoordinationEntryEvent;
+import com.excelian.mache.observable.coordination.CoordinationEntryEvent;
 import com.google.gson.Gson;
 import com.excelian.mache.events.BaseCoordinationEntryEventProducer;
 import org.slf4j.Logger;
