+++ core-observable/src/main/java/com/excelian/mache/events/integration/RabbitMQEventConsumer.java
@@ -6,7 +6,7 @@ import com.rabbitmq.client.AMQP;
 import com.rabbitmq.client.Channel;
 import com.rabbitmq.client.DefaultConsumer;
 import com.rabbitmq.client.Envelope;
-import com.excelian.mache.coordination.CoordinationEntryEvent;
+import com.excelian.mache.observable.coordination.CoordinationEntryEvent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
