+++ core-observable/src/main/java/com/excelian/mache/observable/MapEventListener.java
@@ -1,0 +1,7 @@
+package com.excelian.mache.observable;
+
+import com.excelian.mache.observable.coordination.CoordinationEntryEvent;
+
+public interface MapEventListener {
+    void send(CoordinationEntryEvent<?> event);
+}
