+++ core-observable/src/main/java/com/excelian/mache/observable/coordination/RemoteCacheEntryInvalidateListener.java
@@ -1,4 +1,4 @@
-package com.excelian.mache.coordination;
+package com.excelian.mache.observable.coordination;
 
 import javax.cache.event.CacheEntryListenerException;
