+++ core-observable/src/main/java/com/excelian/mache/observable/coordination/RemoteCacheEntryListener.java
@@ -1,4 +1,4 @@
-package com.excelian.mache.coordination;
+package com.excelian.mache.observable.coordination;
 
 public interface RemoteCacheEntryListener extends RemoteCacheEntryRemovedListener, RemoteCacheEntryInvalidateListener, RemoteCacheEntryCreatedListener, RemoteCacheEntryUpdatedListener {
