+++ core-observable/src/main/resources/META-INF/services/com.excelian.mache.observable.builder.MessagingProvisioner
